Crud Function fror car parts

+++ abc_car_traders/Parts.cs
@@ -16,7 +16,8 @@ namespace abc_car_traders
     public partial class Parts : Form
     {
         CatParts parts = new CatParts();
-
+        AppClass.MyComClass comClass = new AppClass.MyComClass();
+        string AddMessage = "Pleas Fill all fields";
 
         public Parts()
         {
@@ -29,34 +30,65 @@ namespace abc_car_traders
 
         }
 
-        private void modelBox_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            parts.carModel = modelBox.Text;
-        }
-
         private void priceBox_TextChanged(object sender, EventArgs e)
         {
-            parts.price = Convert.ToDecimal(priceBox.Text);
+
+            if (int.TryParse(priceBox.Text, out int price))
+            {
+                parts.price = price;
+            }
         }
 
         private void qtyBox_TextChanged(object sender, EventArgs e)
         {
-            parts.qty = Convert.ToInt32(qtyBox.Text);
+            if (int.TryParse(qtyBox.Text, out int qty))
+            {
+                parts.qty = qty;
+            }
         }
 
-        private void button1_Click(object sender, EventArgs e)
+        private void saveBtn_Click(object sender, EventArgs e)
         {
-            parts.save();
+            if (IsNotValid())
+            {
+                MessageBox.Show(AddMessage);
+            }
+            else
+            {
+                parts.save();
+                parts.view();
+                clearFields();
+            }
+          
         }
 
         private void updateBtn_Click(object sender, EventArgs e)
         {
-            parts.Update();
+            if (IsNotValid())
+            {
+                MessageBox.Show(AddMessage);
+            }
+            else
+            {
+                parts.Update();
+                parts.view();
+                clearFields();
+            }
+            
         }
 
         private void deletBtn_Click(object sender, EventArgs e)
         {
-            parts.delete();
+            if (IsNotValid())
+            {
+                MessageBox.Show(AddMessage);
+            }
+            else
+            {
+                parts.delete();
+                parts.view();
+                clearFields();
+            }
         }
 
         private void Parts_Load(object sender, EventArgs e)
@@ -73,6 +105,23 @@ namespace abc_car_traders
             modelBox.Text = loadDataTable.Rows[rowIndex].Cells[2].Value.ToString();
             priceBox.Text = loadDataTable.Rows[rowIndex].Cells[3].Value.ToString();
             qtyBox.Text = loadDataTable.Rows[rowIndex].Cells[4].Value.ToString();
+
+        }
+
+        private void modelBox_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            parts.carmodel = modelBox.Text;
+        }
+        private void clearFields()
+        {
+            comClass.clearcontroles(nameBox, modelBox, priceBox, qtyBox);
+
+        }
+        private bool IsNotValid()
+        {
+           
+                return comClass.CheckValidateFields(nameBox, modelBox, priceBox, qtyBox); 
+
         }
     }
 }
